Add recovery password feature using PHPMailer
PHPMailer has been added for better email composition and sending.

passolvidada.php is a PHP Web Page that has a form to input your email
and, if submmited, will send an email with the steps needed to recover,
te password. (Missing the database part and the change password page.)

config.php and controlador.php has been changed to comply with the new
additions.